introduce hypervisor.GetBoxInfo
